tools: deduplicate FirmwareDescription helpers
PebbleFirmwareBinaryInfo and the FirmwareDescription struct insertion
existed in two diverged copies: tools/fw_binary_info.py +
tools/insert_firmware_descr.py (manual CLIs) and a vendored copy in
pebble-commander's imaging command, which is the live path for boards
without pblboot (asterix gets the struct prepended at PULSE flash
time).

Make pebble.commander.util.fw_binary_info the canonical
implementation, next to the vendored stm32_crc it depends on, so the
package stays self-contained. The tools/ scripts become thin shims:
fw_binary_info.py subclasses to keep its .elf input support (which
needs the repo-local binutils helper), and insert_firmware_descr.py
keeps its CLI. Output is byte-identical to the previous
implementations.
